+++ CHANGES.md
@@ -41,6 +41,7 @@ Changes:
 - Clean up `haw` names #1582
 - Retrieve lower court information in `nm` #1569
 - Retrieve lower court information in `or` #1569
+- Retrieve lower court information in `sc` #1569
 
 Fixes:
 - Fix connappct #1580